Driving march 31st Updates (#15)
We are going to merge in the working code, hopefully allowing all of the
packages to be a starting point, then start transitioning to ros. The
ros environment will be a full git submodule, with all of the packages
in the local file system including both sets of packages. When we
eventually fix and merge them, we will be able to have the ros packages
override but for now, separate for RSP

---------

Co-authored-by: Tanner <tannergladson@gmail.com>
Co-authored-by: Tanner Gladson <99154505+Tanner-Gladson@users.noreply.github.com>
Co-authored-by: Travis Chen <travis.yu.chen@gmail.com>
Co-authored-by: hejonathan <68175370+hejonathan@users.noreply.github.com>
Co-authored-by: Travis Yu Chen <tchen128@jh.edu>
Co-authored-by: Travis Yu Chen <147548360+TravdaTrav@users.noreply.github.com>